model changes

+++ services/FacebookAccountKit.js
@@ -82,25 +82,46 @@ module.exports = {
 
     signup: async params =>
     {
-        let user, phone;
+        let user, phone, address, motorcycle;
 
-        phone = await strapi.plugins[ 'facebook-account-kit' ].models.phone.create( params.phone );
+        // console.log( strapi );
+        // console.log( strapi.api[ 'address' ] );
+
 
         try
         {
+            address = await strapi.models.address.create( params.address );
+            motorcycle = await strapi.models.motorcycle.create( params.motorcycles[ 0 ] );
+            phone = await strapi.plugins[ 'facebook-account-kit' ].models.phone.create( params.phone );
+            
             user = await strapi.plugins[ 'users-permissions' ].services.user.add(
                 {
                     ...params,
+                    username: params.username ? params.username : params.name,
+                    phone: phone.id,
+                    address: address.id,
+                    motorcycles: [ motorcycle.id ]
                 } );
             return user;
         } catch ( error )
         {
+            console.log( "ERRO CRIANDO" );
             console.log( error );
             try
             {
-                phone = await strapi.plugins[ 'facebook-account-kit' ].models.phone.deleteOne( phone.id );
+                if ( address )
+                    await strapi.models.address.deleteOne( address );
+                if ( motorcycle )
+                    await strapi.models.motorcycle.deleteOne( motorcycle );
+                if ( phone )
+                    await strapi.plugins[ 'facebook-account-kit' ].models.phone.deleteOne( phone.id );
+                if ( user )
+                    await strapi.plugins[ 'users-permissions' ].services.user.remove( user );
+
+
             } catch ( error )
             {
+                console.log( "ERRO REMOVENDO" );
                 console.log( error );
             }
         }
